backend/regulatory_analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.10.0
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
